Move extractor to separate file
