+++ src/sutta_processor/application/domain_models/pali_canon/base.py
@@ -1,6 +1,6 @@
 import logging
 from pathlib import Path
-from typing import Dict, List, Tuple
+from typing import Dict, Tuple
 
 import attr
 from lxml import etree
@@ -13,6 +13,8 @@ from sutta_processor.application.value_objects.uid import (
 )
 from sutta_processor.application.value_objects.verse import PaliVerse
 
+from .extractors import PaliHtmlExtractor
+
 log = logging.getLogger(__name__)
 
 
@@ -21,36 +23,6 @@ class PaliVersus:
     ms_id: PaliMsId
     msdiv_id: PaliMsDivId
     verse: PaliVerse
-
-
-class PaliHtmlExtractor:
-    @classmethod
-    def get_crumb(cls, page: _ElementTree) -> PaliCrumb:
-        last_href: _Element = page.xpath("//CRUMBS/a")[-1]
-        pali_type = PaliCrumb(last_href.get("href"))
-        return pali_type
-
-    @classmethod
-    def get_paragraphs(cls, page: _ElementTree) -> List[_Element]:
-        return page.xpath("//body/p")
-
-    @classmethod
-    def get_ms_msdiv(cls, paragraph: _Element) -> Tuple[PaliMsId, PaliMsDivId]:
-        a_ms = paragraph.xpath("./a[@class='ms']")[0]
-        ms_id = PaliMsId(a_ms.get("id", ""))
-        msdiv_id = PaliMsDivId("")
-        try:
-            a_msdiv = paragraph.xpath("./a[@class='msdiv']")[0]
-            msdiv_id = PaliMsDivId(a_msdiv.get("id", ""))
-        except IndexError:
-            log.debug("No msdiv if for ms: '%s'", ms_id)
-        return ms_id, msdiv_id
-
-    @classmethod
-    def get_verse(cls, paragraph: _Element) -> PaliVerse:
-        text = paragraph.xpath("./text()")[0]
-        versus = PaliVerse(text)
-        return versus
 
 
 @attr.s(frozen=True, auto_attribs=True)
